+++ braille.txt
@@ -1,3 +1,3 @@
-0.0.000.
-0...0..0
-00..0...
+0.0.0.0.0.
+00.00.0..0
+....0.0.0.
